+++ src/CG.Linq.EFCore/ServiceCollectionExtensions.cs
@@ -24,7 +24,6 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <typeparam name="TOptions">The type of associated repository options.</typeparam>
         /// <param name="serviceCollection">The service collection to use for 
         /// the operation.</param>
-        /// <param name="dataProtector">The data protector to use for the operation.</param>
         /// <param name="configuration">The configuration to use for the operation.</param>
         /// <param name="serviceLifetime">The service lifetime to use for the operation.</param>
         /// <returns>The value of the <paramref name="serviceCollection"/>
@@ -33,19 +32,16 @@ namespace Microsoft.Extensions.DependencyInjection
         /// or more arguments are invalid, or missing.</exception>
         public static IServiceCollection AddEFCore<TOptions>(
             this IServiceCollection serviceCollection,
-            IDataProtector dataProtector,
             IConfiguration configuration,
             ServiceLifetime serviceLifetime = ServiceLifetime.Scoped
             ) where TOptions : EFCoreRepositoryOptions, new()
         {
             // Validate the parameters before attempting to use them.
             Guard.Instance().ThrowIfNull(serviceCollection, nameof(serviceCollection))
-                .ThrowIfNull(dataProtector, nameof(dataProtector))
                 .ThrowIfNull(configuration, nameof(configuration));
 
             // Register the repository options.
             serviceCollection.ConfigureOptions<TOptions>(
-                dataProtector,
                 configuration
                 );